+++ flask_app/search/__init__.py
@@ -1,0 +1,1 @@
+from .search import get_orm_query_from_search_string