+++ flask_app/search/exceptions.py
@@ -1,0 +1,11 @@
+class SearchException(Exception):
+    pass
+
+class SearchSyntaxError(SearchException):
+    pass
+
+class UnknownField(SearchSyntaxError):
+    pass
+
+class UnknownOperator(SearchSyntaxError):
+    pass