+++ tests/test_warnings.py
@@ -1,7 +1,7 @@
 import pytest
 
 from .utils import raises_not_found
-from backslash.test import Test as TestType
+from backslash import test
 
 
 def test_add_warnings(warning_container, filename, lineno, message, timestamp):
@@ -13,7 +13,7 @@ def test_add_warnings(warning_container, filename, lineno, message, timestamp):
     assert fetched.lineno == lineno
     assert fetched.timestamp == timestamp
     assert fetched.filename == filename
-    if isinstance(warning_container, TestType):
+    if isinstance(warning_container, test.Test):
         assert warning_container.get_session().num_test_warnings == 1
         assert warning_container.get_session().num_warnings == 0