+++ webapp/app/components/keyboard-shortcuts/component.js
@@ -11,6 +11,7 @@ let _keys = [
     {key: 'h', action: 'toggle_human_times', description: 'Toggles human-readable times'},
     {key: 'a', action: 'filter_none', description: 'Show all entities'},
     {key: 'f', action: 'filter_only_failed', description: 'Hide all entities except failed'},
+    {key: 'c', action: 'toggle_inline_comment', description: 'Toggle comment preview for items'},
     {key: 'esc', action: 'close_boxes_or_home'},
     {key: 'ctrl+s', action: 'goto_sessions', description: 'Jump to Sessions view'},
     {key: 'ctrl+u', action: 'goto_users', description: 'Jump to Users view'},
@@ -41,6 +42,7 @@ export default Ember.Component.extend(KeyboardShortcuts, {
     keys: _keys,
 
     api: Ember.inject.service(),
+    display: Ember.inject.service(),
 
 
     search(query, sync_callback, async_callback) {
@@ -117,9 +119,7 @@ export default Ember.Component.extend(KeyboardShortcuts, {
         },
 
         toggle_human_times() {
-            this._do_if_in(_FILTERABLE_VIEWS, function(controller) {
-                controller.toggleProperty('humanize_times');
-            });
+	    this.get('display').toggleProperty('humanize_times');
         },
 
 
@@ -134,6 +134,10 @@ export default Ember.Component.extend(KeyboardShortcuts, {
                 controller.filter_none();
             });
         },
+
+	toggle_inline_comment() {
+	    this.get('display').toggleProperty('comments_expanded');
+	},
 
         open_quick_search() {
             let self = this;