+++ webapp/app/components/test-item/component.js
@@ -1,31 +1,18 @@
 import Ember from 'ember';
 
 export default Ember.Component.extend({
+    display: Ember.inject.service(),
     attributeBindings: ['href'],
     tagName: 'a',
     classNames: ['item', 'test', 'clickable'],
 
-    classNameBindings: ['status', 'has_any_error:unsuccessful'],
+    classNameBindings: ['test.computed_status', 'has_any_error:unsuccessful'],
 
     test: Ember.computed.oneWay('item'),
     session_model: null,
 
-    status: function() {
-        let status = this.get('test.status').toLowerCase();
-        if (status === 'failure' || status === 'error') {
-            status = 'failed';
-        }
-        return status;
-    }.property('test.status'),
-
-    is_running: function() {
-	let session = this.get('session_model');
-	if (session && session.get('is_abandoned')) {
-	    return false;
-	}
-	return this.get('status') === 'running';
-    }.property('session_model', 'status'),
-
+    is_running: Ember.computed.equal('test.computed_status', 'running'),
+    
     href: function() {
 	let returned = `/#/sessions/${this.get('test.session_display_id')}/tests/${this.get('test.display_id')}`;
 	if (this.get('test.has_any_error')) {