+++ webapp/tests/integration/components/item-comment-overview/component-test.js
@@ -1,0 +1,16 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('item-comment-overview', 'Integration | Component | item comment overview', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{item-comment-overview}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+});